requirements.txt
fastapi
uvicorn[standard]
python-multipart
mediapipe
opencv-python-headless
numpy
pillow
